% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\docType{data}
\name{default_AAR_params}
\alias{default_AAR_params}
\title{Default AAR kinetic parameters}
\format{
A named list with four elements:
\describe{
\item{Ae}{Pre-exponential (frequency) factor: \code{exp(42.12)}.}
\item{Ea}{Activation energy in kcal/mol: \code{31.5}.}
\item{R}{Gas constant in kcal K\eqn{^{-1}} mol\eqn{^{-1}}: \code{0.001987}.}
\item{x}{Power-law exponent for D/L linearization: \code{3}.}
}
}
\usage{
default_AAR_params
}
\description{
A named list of Arrhenius and power-law constants used as defaults
throughout the package. Values are from Kaufman (2000) for isoleucine
epimerization and have not yet been calibrated to site-specific data.
}
\keyword{datasets}
