% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{plot_mcmc_chains}
\alias{plot_mcmc_chains}
\title{Trace plots for MCMC output}
\usage{
plot_mcmc_chains(mcmc_out, burnin = 1000)
}
\arguments{
\item{mcmc_out}{List returned by \code{\link{run_mcmc}}.}

\item{burnin}{Integer. Burn-in iterations to shade (not removed). Default
\code{1000}.}
}
\value{
A ggplot object.
}
\description{
Trace plots for MCMC output
}
