% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{plotInputs}
\alias{plotInputs}
\title{Plot temperature inputs for a damped sediment record}
\usage{
plotInputs(dampened)
}
\arguments{
\item{dampened}{Data frame returned by \code{\link{dampSedTemps}}.}
}
\value{
A ggplot object.
}
\description{
Convenience plotting function for the output of \code{\link{dampSedTemps}}.
Pivots the data frame to long format and draws air and sediment temperature
histories.
}
